requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
google-generativeai
sentence-transformers
faiss-cpu
pandas
numpy
python-dotenv
